factor out common code to be reused
