+++ router/gin/storage.go
@@ -1,0 +1,27 @@
+package gin
+
+import (
+	"context"
+
+	krakendrate "github.com/krakendio/krakend-ratelimit/v3"
+	"github.com/krakendio/krakend-ratelimit/v3/router"
+)
+
+func StoreFromCfg(cfg router.Config) krakendrate.LimiterStore {
+	ctx := context.Background()
+	var storeBackend krakendrate.Backend
+	if cfg.NumShards > 1 {
+		storeBackend = krakendrate.NewShardedBackend(
+			ctx,
+			cfg.NumShards,
+			cfg.TTL,
+			krakendrate.PseudoFNV64a,
+			krakendrate.MemoryBackendBuilder,
+		)
+	} else {
+		storeBackend = krakendrate.MemoryBackendBuilder(ctx, cfg.TTL)
+	}
+
+	return krakendrate.NewLimiterStore(cfg.ClientMaxRate, int(cfg.ClientCapacity),
+		storeBackend)
+}
